+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 2.0 preview work in progress
 ------------------------------------
 - enh #1: Remove engine\config.classes.php, update classes autoloading (AzatGaliev)
 - enh #3: Changed the style code in all files (VladimirShestakov)
+- bug: Updated algorithm of working embedded values ​​in the class \Boolive\values\Values. Removed references to property _value
 - new #4: Add CHANGELOG.md file (AzatGaliev)
 - new #10: Update all objects in "library/basic" with added contents, members and simple objects (VladimirShestakov)
 - new #14: Added layout widget "boolive" in the "library/layouts" and "interfaces/html/body" (VladimirShestakov)
@@ -16,6 +17,7 @@ Version 2.0 preview work in progress
 - new #20: Added widget for view Page (VladimirShestakov)
 - new #32: Added widget for view Part (VladimirShestakov)
 - new #33: Added widget Focuser (VladimirShestakov)
+- new #35: Added widget PageNavigation in the basic package and layout boolive for widget the part (VladimirShestakov)
 - new: Added object the Option for automatic views (VladimirShestakov)
 - bug: Fix some paths in code for compatibility with UNIX (AzatGaliev)
 - new: Added top menu and sidebar menu in the boolive layout (VladimirShestakov)